test(jit): reorganize/cleanup the tests a bit
